+++ integrations/amazon_bedrock/tests/test_chat_generator.py
@@ -22,9 +22,18 @@ MODELS_TO_TEST_WITH_TOOLS = [
 ]
 
 # so far we've discovered these models support streaming and tool use
-STREAMING_TOOL_MODELS = ["anthropic.claude-3-5-sonnet-20240620-v1:0", "cohere.command-r-plus-v1:0"]
+STREAMING_TOOL_MODELS = [
+    "anthropic.claude-3-5-sonnet-20240620-v1:0",
+    "us.anthropic.claude-sonnet-4-20250514-v1:0",
+    "cohere.command-r-plus-v1:0",
+]
 
 MODELS_TO_TEST_WITH_IMAGE_INPUT = [
+    "us.anthropic.claude-sonnet-4-20250514-v1:0",
+]
+
+MODELS_TO_TEST_WITH_THINKING = [
+    "us.anthropic.claude-3-7-sonnet-20250219-v1:0",
     "us.anthropic.claude-sonnet-4-20250514-v1:0",
 ]
 
@@ -399,6 +408,160 @@ class TestAmazonBedrockChatGeneratorInference:
         assert len(final_message.text) > 0
         assert "paris" in final_message.text.lower()
         assert "berlin" in final_message.text.lower()
+
+    @pytest.mark.parametrize("model_name", MODELS_TO_TEST_WITH_THINKING)
+    def test_live_run_with_tool_call_and_thinking(self, model_name, tools):
+        initial_messages = [ChatMessage.from_user("What's the weather like in Paris?")]
+        component = AmazonBedrockChatGenerator(
+            model=model_name,
+            tools=tools,
+            generation_kwargs={
+                "maxTokens": 8192,
+                "thinking": {
+                    "type": "enabled",
+                    "budget_tokens": 1024,
+                },
+            },
+        )
+        results = component.run(messages=initial_messages)
+
+        assert len(results["replies"]) > 0, "No replies received"
+
+        # Find the message with tool calls
+        tool_call_message = None
+        for message in results["replies"]:
+            if message.tool_calls:
+                tool_call_message = message
+                break
+
+        assert tool_call_message is not None, "No message with tool call found"
+        assert isinstance(tool_call_message, ChatMessage), "Tool message is not a ChatMessage instance"
+        assert ChatMessage.is_from(tool_call_message, ChatRole.ASSISTANT), "Tool message is not from the assistant"
+
+        tool_calls = tool_call_message.tool_calls
+        assert len(tool_calls) == 1
+        assert tool_calls[0].id, "Tool call does not contain value for 'id' key"
+        assert tool_calls[0].tool_name == "weather"
+        assert tool_calls[0].arguments["city"] == "Paris"
+        assert tool_call_message.meta["finish_reason"] == "tool_use"
+
+        # Mock the response we'd get from ToolInvoker
+        tool_result_messages = [
+            ChatMessage.from_tool(tool_result="22° C", origin=tool_call) for tool_call in tool_calls
+        ]
+
+        new_messages = [*initial_messages, tool_call_message, *tool_result_messages]
+        results = component.run(new_messages)
+
+        assert len(results["replies"]) == 1
+        final_message = results["replies"][0]
+        assert not final_message.tool_call
+        assert len(final_message.text) > 0
+        assert "paris" in final_message.text.lower()
+
+    @pytest.mark.parametrize("model_name", MODELS_TO_TEST_WITH_THINKING)
+    def test_live_run_with_tool_call_and_thinking_streaming(self, model_name, tools):
+        initial_messages = [ChatMessage.from_user("What's the weather like in Paris?")]
+        component = AmazonBedrockChatGenerator(
+            model=model_name,
+            tools=tools,
+            generation_kwargs={
+                "maxTokens": 8192,
+                "thinking": {
+                    "type": "enabled",
+                    "budget_tokens": 1024,
+                },
+            },
+            streaming_callback=print_streaming_chunk,
+        )
+        results = component.run(messages=initial_messages)
+
+        assert len(results["replies"]) > 0, "No replies received"
+
+        # Find the message with tool calls
+        tool_call_message = None
+        for message in results["replies"]:
+            if message.tool_calls:
+                tool_call_message = message
+                break
+
+        assert tool_call_message is not None, "No message with tool call found"
+        assert isinstance(tool_call_message, ChatMessage), "Tool message is not a ChatMessage instance"
+        assert ChatMessage.is_from(tool_call_message, ChatRole.ASSISTANT), "Tool message is not from the assistant"
+
+        tool_calls = tool_call_message.tool_calls
+        assert len(tool_calls) == 1
+        assert tool_calls[0].id, "Tool call does not contain value for 'id' key"
+        assert tool_calls[0].tool_name == "weather"
+        assert tool_calls[0].arguments["city"] == "Paris"
+        assert tool_call_message.meta["finish_reason"] == "tool_use"
+
+        # Mock the response we'd get from ToolInvoker
+        tool_result_messages = [
+            ChatMessage.from_tool(tool_result="22° C", origin=tool_call) for tool_call in tool_calls
+        ]
+
+        new_messages = [*initial_messages, tool_call_message, *tool_result_messages]
+        results = component.run(new_messages)
+
+        assert len(results["replies"]) == 1
+        final_message = results["replies"][0]
+        assert not final_message.tool_call
+        assert len(final_message.text) > 0
+        assert "paris" in final_message.text.lower()
+
+    def test_live_run_with_redacted_thinking(self, tools):
+        # Special prompt to induce redacted thinking
+        initial_messages = [
+            ChatMessage.from_user(
+                "ANTHROPIC_MAGIC_STRING_TRIGGER_REDACTED_THINKING_46C9A13E193C177646C7398A98432ECCCE4C1253D5E2D82641AC0E52CC2876CB"
+            )
+        ]
+        component = AmazonBedrockChatGenerator(
+            # Redacted thinking only happens with Claude 3.7 https://docs.aws.amazon.com/bedrock/latest/userguide/claude-messages-thinking-encryption.html
+            model="us.anthropic.claude-3-7-sonnet-20250219-v1:0",
+            tools=tools,
+            generation_kwargs={
+                "maxTokens": 8192,
+                "thinking": {
+                    "type": "enabled",
+                    "budget_tokens": 1024,
+                },
+            },
+        )
+        results = component.run(messages=initial_messages)
+
+        assert len(results["replies"]) > 0, "No replies received"
+        assert isinstance(
+            results["replies"][0].meta["reasoning_contents"][0]["reasoning_content"]["redacted_content"], bytes
+        )
+
+    def test_live_run_with_redacted_thinking_streaming(self, tools):
+        # Special prompt to induce redacted thinking
+        initial_messages = [
+            ChatMessage.from_user(
+                "ANTHROPIC_MAGIC_STRING_TRIGGER_REDACTED_THINKING_46C9A13E193C177646C7398A98432ECCCE4C1253D5E2D82641AC0E52CC2876CB"
+            )
+        ]
+        component = AmazonBedrockChatGenerator(
+            # Redacted thinking only happens with Claude 3.7 https://docs.aws.amazon.com/bedrock/latest/userguide/claude-messages-thinking-encryption.html
+            model="us.anthropic.claude-3-7-sonnet-20250219-v1:0",
+            tools=tools,
+            generation_kwargs={
+                "maxTokens": 8192,
+                "thinking": {
+                    "type": "enabled",
+                    "budget_tokens": 1024,
+                },
+            },
+            streaming_callback=print_streaming_chunk,
+        )
+        results = component.run(messages=initial_messages)
+
+        assert len(results["replies"]) > 0, "No replies received"
+        assert isinstance(
+            results["replies"][0].meta["reasoning_contents"][0]["reasoning_content"]["redacted_content"], bytes
+        )
 
     @pytest.mark.parametrize("model_name", STREAMING_TOOL_MODELS)
     def test_live_run_with_multi_tool_calls_streaming(self, model_name, tools):